separate out firestore transaction integration tests. add PhoneNumberRepository.
