+++ spring-cloud-gcp-data-firestore/src/test/java/com/google/cloud/spring/data/firestore/entities/PhoneNumberRepository.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.cloud.spring.data.firestore.entities;
+
+import com.google.cloud.spring.data.firestore.FirestoreReactiveRepository;
+
+// temp: added for testing to bypass "An ID property was required but does not exist for the type" error
+public interface PhoneNumberRepository extends FirestoreReactiveRepository<PhoneNumber> {}
